Do what the "pre" comments say.
We don't clone the remote repository anymore. Instead we create a new repository and link it to
the remote repository, then pull the gh-pages branch. We deploy the site and check everything back in
and push back to gh-pages.

This way we don't have to pull all the source to deploy the site.